front-end/src/app/shared/desplegable-tipo-movimiento/tipo-movimiento.service.spec.ts
import { TestBed } from '@angular/core/testing';

import { TipoMovimientoService } from './tipo-movimiento.service';
import { TestModule } from '../../core/mocks/test.module';
import { ApiService } from '../../core/services/restClient/api.service';
import { configureTestSuite } from '../../core/mocks/testing';
import { TipoMovimiento } from '../data-models/tipo-movimiento';
import { of } from 'rxjs';

describe('TipoMovimientoService', () => {
  let apiService: any;
  let service: TipoMovimientoService;

  configureTestSuite(() => {
    apiService = jasmine.createSpyObj('ApiService', ['get']);
    TestBed.configureTestingModule({
      imports: [TestModule],
      providers: [ TipoMovimientoService,
      {provide: ApiService, useValue: apiService}]
    });
    service = TestBed.get(TipoMovimientoService);
  });

  beforeEach(() => {
    apiService.get.calls.reset();
  });

  it('should be created', () => {
    expect(service).toBeTruthy();
  });

  describe('El metodo getAll()', () => {
    it('invoca al metodo get del ApiService', () => {
      service.getAll();

      expect(apiService.get).toHaveBeenCalledWith('tipos-movimiento');
      expect(apiService.get).toHaveBeenCalledTimes(1);
    });

    it('devuelve un observable con los datos obtenidos por el apiservice.get', () => {
      const esperado = of([{} as TipoMovimiento]);
      apiService.get.and.returnValue(esperado);

      const resultado = service.getAll();

      expect(resultado).toEqual(esperado);
    });
  });
});
